Recreate session when making new client
requests has a memory leak from our observation of containers
experiencing OOMs for this sidecar. It appears that the underlying
session is keeping sockets alive. Calling close on the session should
clean things up. We don't want to do this all the time because the point
is to reuse connections in the pool. However, when we expire a client we
can take that opportunity to recreate the session. Over time this should
let us clean up the memory help by the session while having a minimal
impact on connection reuse.